Added correct use of threading for metadata gathering, as well as some more configuration options.
TODO: Add an optional grace period to wait before initiating a metadata search- skipping through 20 songs in a library to play only one in its entirety is wasteful when requests for metadata are made at every track skip.

The plyr return values need to be looked into- seems to not always be returning the correct data.